+++ shelter-project/src/app/dogs/dogs.component.ts
@@ -61,14 +61,16 @@ export class DogsComponent implements OnInit {
   }
 
   delete (id: number,event) {
-    event.stopPropagation()
-    console.log(event)
-    this.animalService.deleteAnimal(id).then(()=>{
-      console.log("SIKERES TORLES")
+    if (confirm('Biztos törölni szeretné?')) {
+      event.stopPropagation()
+      console.log(event)
+      this.animalService.deleteAnimal(id).then(()=>{
+        console.log("SIKERES TORLES")
+        this.loadAnimals()
+      })
+    }else{
       this.loadAnimals()
-    })
-   
-    
+    }
   }
 
 }
